Fix caret reset position bug
The field host renderer was being re-rendered before its parent the block host renderer on value changes. This is because Redux Zero re-renders the calling component of actions, which caused the field host renderer to be rendered with its previous value. Since the formPathToValue performance  degradation is negligible the value is now calculated in the connect function.

Fixes https://github.com/DeReCRUD/de-re-crud/issues/46