feat: expose field labels as renderers
This is the introduction of a new pattern that will most likely be
applied to more of the renderers in the future to make it easier to
create custom implementations. The framework now provides a function
that can be used to simplify rendering labels within field renderers.
This fully backwards compatible with the old pattern to allow for
extreme flexibility.

BREAKING CHANGE: Renderer definitions have been updated